pyproject.toml
[project]
name = "gordon_nemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lightning>2.2.1,<=2.4.0",
    "hydra-core>=1.3.2",
    "nemo-toolkit>=2.1.0",
    "numpy>=2.2.6",
    "jaxtyping>=0.3.2",
    "matplotlib>=3.10.3",
    "braceexpand",
    "cloudpickle>=3.1.1",
    "datasets>=3.6.0",
    "editdistance>=0.8.1",
    "einops>=0.8.1",
    "ffmpeg>=1.4",
    "fiddle",
    "h5py>=3.13.0",
    "huggingface-hub>=0.31.1",
    "ijson>=3.4.0",
    "ipykernel>=6.29.5",
    "jieba>=0.42.1",
    "jiwer>=3.1.0",
    "jupyterlab>=4.4.2",
    # "lhotse>=1.31.0",
    "lhotse>=1.30.0,<1.31.0",
    "librosa>=0.11.0",
    "megatron-core>=0.12.0",
    "nem>=0.1",
    "numba>=0.61.2",
    "onnx>=1.17.0",
    "opencc>=1.1.9",
    "pandas>=2.2.3",
    "pangu>=4.0.6.1",
    "protobuf>=6.30.2",
    "pyannote-core>=5.0.0",
    "pyannote-metrics>=3.2.1",
    "rouge-score>=0.1.2",
    "ruamel-yaml>=0.18.10",
    "sacrebleu>=2.5.1",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "sentencepiece>=0.2.0",
    "setuptools>=80.4.0",
    "sox>=1.5.0",
    "tensorboard>=2.19.0",
    "text-unidecode>=1.3",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "tqdm>=4.67.1",
    "transformers>=4.51.3",
    "webdataset>=0.2.111",
    "wget>=3.2",
    "wrapt>=1.17.2",
]

[tool.setuptools]
packages=["multiblock_regression", "model_pt_test"]
package-dir={"" = "."}
